requirements.txt
# ------------------------------
# Core framework and web server
# ------------------------------
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6

# ------------------------------
# Database
# ------------------------------
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# ------------------------------
# Authentication and security
# ------------------------------
PyJWT==2.8.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
cryptography==41.0.7

# ------------------------------
# Cache and sessions
# ------------------------------
redis<=5.0.1
hiredis==2.2.3

# ------------------------------
# HTTP client and requests
# ------------------------------
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.0

# ------------------------------
# OCR and image processing
# ------------------------------
pytesseract==0.3.10
opencv-python==4.9.0.80
Pillow
pdf2image==1.16.3
easyocr==1.7.0
wand==0.6.13

# ------------------------------
# AI and ML
# ------------------------------
numpy==1.26.4
torch==2.1.1
scikit-learn==1.3.2
transformers==4.36.0
sentence-transformers==5.1.1

# ------------------------------
# Email integration
# ------------------------------
email-validator==2.1.0
imapclient==2.1.0
# ------------------------------
# File processing and monitoring
# ------------------------------
watchdog==3.0.0
aiofiles==23.2.1
python-magic==0.4.27

# ------------------------------
# Vector database and search
# ------------------------------
faiss-cpu==1.7.4
chromadb==0.4.18
pandas==2.1.4

# ------------------------------
# Document processing
# ------------------------------
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
textract==1.6.5

# ------------------------------
# Language detection and NLP
# ------------------------------
langdetect==1.0.9
spacy==3.7.2
nltk==3.8.1

# ------------------------------
# Email and notifications
# ------------------------------
jinja2==3.1.2
aiosmtplib==3.0.1

# ------------------------------
# Data integrations
# ------------------------------
exchangelib==5.2.0
Office365-REST-Python-Client==2.4.4
google-api-python-client==2.108.0
google-auth==2.24.0
google-auth-oauthlib==1.0.0

# ------------------------------
# Async and background tasks
# ------------------------------
celery[redis]==5.3.4
asyncio-mqtt==0.14.0

# ------------------------------
# File handling and utilities
# ------------------------------
python-decouple==3.8
pathlib2==2.3.7
python-dotenv==1.0.0
pydantic-settings==2.1.0

# ------------------------------
# Monitoring and metrics
# ------------------------------
prometheus-client==0.19.0
structlog==23.2.0

# ------------------------------
# Development and testing
# ------------------------------
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# ------------------------------
# WebSocket support
# ------------------------------
websockets==12.0
python-socketio==5.10.0

# ------------------------------
# Date and time handling
# ------------------------------
python-dateutil==2.8.2
pytz==2023.3

# ------------------------------
# JSON and data serialization
# ------------------------------
orjson==3.9.10
msgpack==1.0.7

# ------------------------------
# Malayalam language support
# ------------------------------
indic-nlp-library==0.81

# ------------------------------
# WhatsApp integration
# ------------------------------
whatsapp-api-client-python==0.0.49

# ------------------------------
# Cloud storage
# ------------------------------
boto3==1.34.0
azure-storage-blob==12.19.0
google-cloud-storage==2.10.0

# ------------------------------
# Backup and archival
# ------------------------------
py7zr==0.20.8
zipfile36==0.1.3

# ------------------------------
# Rate limiting
# ------------------------------
slowapi==0.1.9
limits==3.7.0

# ------------------------------
# CORS support
# ------------------------------
fastapi-cors==0.0.6

# ------------------------------
# Health checks
# ------------------------------
fastapi-health==0.4.0

# ------------------------------
# Additional dependencies for new services
# ------------------------------
schedule==1.2.0
psutil==5.9.6
pyotp==2.9.0
qrcode[pil]==7.4.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
kaleido==0.2.1
